if the key is not present in the hashmap at all, we return false
i know this is kind of inefficient to then find the key again in the right bucket, but i'm not sure what to do here. i'll come back and refactor later if i have better idea

+++ createHashMap.mjs
@@ -46,7 +46,7 @@ export default function createHashMap() {
     const hashCode = hash(key);
     const bucket = buckets[hashCode];
 
-    if (bucket.isEmpty()) return false;
+    if (bucket.isEmpty() || !has(key)) return false;
     bucket.removeNode(key);
 
     return true;
